avance PUT

+++ src/main/java/com/ubb/proyecto/service/InfoContactService.java
@@ -20,13 +20,19 @@ public class InfoContactService {
         return infoContactRepository.findAll();
     }
 
-    public Optional<InfoContact> getInfoContactById(Integer id) {
+    /*public Optional<InfoContact> getInfoContactById(Integer id) {
         return infoContactRepository.findById(id);
+    }*/
+    public Optional<InfoContact> getSingleInfoContact() {
+        return infoContactRepository.findFirstByOrderByIdContactAsc();
     }
 
     public InfoContact saveInfoContact(InfoContact infoContact/*, Integer id_user*/){
         /*Usuario usuario = usuarioService.getUsuariosById(id_user).orElseThrow(() -> new RuntimeException("Usuario no encontrado"));
         infoContact.setcambio(usuario);*/
+        if (infoContact.getCorreo() == null || infoContact.getTelefono() == null) {
+            throw new IllegalArgumentException("Los campos 'correo' y 'telefono' son obligatorios.");
+        }
         return infoContactRepository.save(infoContact);
     }
 
@@ -43,7 +49,7 @@ public class InfoContactService {
     }*/
 
    //Solo actualiza correo?
-    public InfoContact updateInfoContact(Integer id, InfoContact InfoContactDetails) {
+    /*public InfoContact updateInfoContact(Integer id, InfoContact InfoContactDetails) {
         InfoContact infoContact = infoContactRepository.findById(id)
                 .orElseThrow(() -> new RuntimeException("Contacto no encontrado"));
 
@@ -51,9 +57,26 @@ public class InfoContactService {
 
 
         return infoContactRepository.save(infoContact);
+    }*/
+
+    public InfoContact updateInfoContact(Integer id, InfoContact newInfoContact) {
+        InfoContact existingContact = infoContactRepository.findById(id).orElseThrow(() -> new RuntimeException("InfoContact no encontrado."));
+        existingContact.setCorreo(newInfoContact.getCorreo());
+        existingContact.setTelefono(newInfoContact.getTelefono());
+        existingContact.setInstagram(newInfoContact.getInstagram());
+        existingContact.setFacebook(newInfoContact.getFacebook());
+        //existingContact.setUpdated_by(newInfoContact.getUpdated_by()); comentaod xq está lo de abajo
+
+        if (newInfoContact.getUpdated_by() != null) {
+            existingContact.setUpdated_by(newInfoContact.getUpdated_by());
+        }
+        return infoContactRepository.save(existingContact);
     }
 
     public void deleteInfoContact(Integer id) {
+        if (!infoContactRepository.existsById(id)) {
+            throw new RuntimeException("InfoContact no encontrado.");
+        }
         infoContactRepository.deleteById(id);
     }
